requirements.txt
numpy==1.26.0
einops
gradio
omegaconf
matplotlib
accelerate>=0.22.0
torch==2.2.0 --index-url https://download.pytorch.org/whl/cu121
torchdiffeq>=0.2.3
diffusers==0.26.3
huggingface_hub==0.25.0
xformers==0.0.24 
